requirements.txt
httpx
pydantic
python-dotenv
PyYAML
pypdf
reportlab
Pillow
img2pdf
tenacity
openpyxl
